+++ fold_node/src/datafold_node/sample_manager.rs
@@ -131,18 +131,5 @@ impl SampleManager {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::SampleManager;
 
-    #[tokio::test]
-    async fn sample_manager_loads_schemas() {
-        let manager = SampleManager::new().await.expect("failed to load samples");
-        let schemas = manager.list_schema_samples();
-        assert!(schemas.contains(&"UserProfile".to_string()));
-        assert!(schemas.contains(&"ProductCatalog".to_string()));
-        assert!(schemas.contains(&"UserProfileView".to_string()));
-        assert!(schemas.contains(&"BlogPostSummary".to_string()));
-    }
-}
