+++ fold_node/src/datafold_node/schema_routes.rs
@@ -1,5 +1,3 @@
-#[cfg(test)]
-use super::sample_manager::SampleManager;
 use super::http_server::AppState;
 use crate::schema::Schema;
 use actix_web::{web, HttpResponse, Responder};
@@ -97,26 +95,5 @@ pub async fn load_available_schema_route(path: web::Path<String>, state: web::Da
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::datafold_node::{DataFoldNode, config::NodeConfig};
-    use actix_web::web;
-    use tempfile::tempdir;
 
-    #[tokio::test]
-    async fn list_schemas_empty() {
-        let dir = tempdir().unwrap();
-        let config = NodeConfig::new(dir.path().to_path_buf());
-        let node = DataFoldNode::new(config).unwrap();
-        let state = web::Data::new(super::super::http_server::AppState {
-            node: std::sync::Arc::new(tokio::sync::Mutex::new(node)),
-            sample_manager: SampleManager { schemas: Default::default(), queries: Default::default(), mutations: Default::default() }
-        });
-        use actix_web::test;
-        let req = test::TestRequest::default().to_http_request();
-        let resp = list_schemas(state).await.respond_to(&req);
-        assert_eq!(resp.status(), 200);
-    }
-}
